Added Boss and Move classes
+ Boss Class & framework
+ Move Class & framework

* Changed Player.takeDamage() to use an int as its argument
* Changed Enemy.attackPlayer to send Player.takeDamage the weapon damage instead of the raw weapon object

+++ src/com/java/Weapon.java
@@ -18,7 +18,7 @@ public class Weapon {
         return this.name;
     }
 
-    public float getDmg(){
+    public int getDmg(){
         return this.dmg;
     }
